Cambie de double a int en resolver

+++ estudiantes/src/main/java/aed/Estudiante.java
@@ -7,7 +7,7 @@ public class Estudiante {
     int fila;
     int columna;
     int correctas;
-    double nota;
+    int nota;
     boolean entrego;
     boolean sospechoso;
     Examen examen;
